Refactor prompt editing hooks
